onboard: remove gsettings inherit
The gsettings bbclass adds postinst hooks to regenerate the glib-2
schemas files, on the assumption that the inheriting recipe actually
contains new schemas. Onboard does not contain such things, which causes
the recipe postinst to fail, in cases where no other
gsettings-containing packages are installed to the system.

Remove the unneeded gsettings bbclass.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>